refactor: move implement proposal action methods to proposal-actions.service
