+++ src/proposals/proposals.module.ts
@@ -1,8 +1,6 @@
 import { forwardRef, Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { CommentsModule } from "../comments/comments.module";
-import { EventsModule } from "../events/events.module";
-import { GroupRolesModule } from "../groups/group-roles/group-roles.module";
 import { GroupsModule } from "../groups/groups.module";
 import { ImagesModule } from "../images/images.module";
 import { VotesModule } from "../votes/votes.module";
@@ -19,8 +17,6 @@ import { ProposalsService } from "./proposals.service";
     forwardRef(() => CommentsModule),
     forwardRef(() => ProposalActionsModule),
     forwardRef(() => VotesModule),
-    EventsModule,
-    GroupRolesModule,
     GroupsModule,
     ImagesModule,
     ProposalActionEventsModule,
